requirements.txt
alembic==1.7.4
Babel==2.9.1
blinker==1.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dominate==2.6.0
elasticsearch==7.15.1
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.5
MarkupSafe==1.1.1
PyJWT==2.3.0
PyMySQL==1.0.2
pytest-html==3.1.1
python-dateutil==2.8.1
python-dotenv==0.19.1
pytz==2021.1
requests==2.25.0
six==1.15.0
SQLAlchemy==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
wxPython==4.1.1
